+++ R/download_s2.R
@@ -8,7 +8,7 @@
 #' @param S2_items list. S2 items already downloaded
 #' @param offset numeric. S2 offset value
 #' @param offset_B2 boolean.
-#' @param collection character.
+#' @param stac_info list.
 #' @param corr_BRF boolean.
 #' @param siteName character. name of the study site
 #' @param crs_target numeric.
@@ -23,24 +23,17 @@
 #'
 download_s2 <- function(aoi, raster_dir, collection_path, iChar, resolution,
                         S2_items = NULL, offset = 1000, offset_B2 = F,
-                        collection = 'sentinel-2-l2a', corr_BRF = F,
+                        stac_info, corr_BRF = F,
                         siteName = NULL, crs_target = NULL, writeoutput = T,
                         bands2correct = c('B8A', 'B11', 'B12')){
 
   item_collection <- readRDS(file = collection_path)
-  if (collection == 'sentinel-2-l2a'){
-    baseline <- lapply(lapply(item_collection$features,'[[','properties'),
-                       '[[', 's2:processing_baseline')
-    asset_cloud <- 'SCL'
-  } else if (collection == 'sentinel2-l2a-sen2lasrc'){
-    baseline <- lapply(lapply(lapply(item_collection$features,'[[','properties'),
-                              '[[', 'processing:software'),
-                       '[[', 'sen2lasrc')
-    asset_cloud <- 'CLM'
-  }
+  asset_cloud <- get_cloud_asset(stac_info = stac_info)
+  baseline <- get_s2_baseline(stac_info, item_collection)
 
   # get bounding box
-  asset_names <- c('B02', 'B03', 'B04', 'B05', 'B06', 'B07', 'B08', 'B8A', 'B11', 'B12')
+  asset_names <- c('B02', 'B03', 'B04', 'B05', 'B06',
+                   'B07', 'B08', 'B8A', 'B11', 'B12')
   asset_names_list <- list_assets(S2_items = S2_items,
                                   asset_names = asset_names,
                                   dateAcqs = item_collection$acquisitionDate)
@@ -48,7 +41,7 @@ download_s2 <- function(aoi, raster_dir, collection_path, iChar, resolution,
   S2_items_update <- mapply(FUN = get_asset_terra,
                             item = item_collection$features,
                             asset_names = asset_names_list,
-                            MoreArgs = list(collection = collection,
+                            MoreArgs = list(collection = stac_info$collection,
                                             aoi = aoi,
                                             crs_target = crs_target),
                             SIMPLIFY = F)
@@ -57,12 +50,15 @@ download_s2 <- function(aoi, raster_dir, collection_path, iChar, resolution,
   for (i in seq_len(length(baseline))){
     if (as.numeric(baseline[[i]])>=4 & asset_cloud == 'SCL')
       S2_items_update[[i]] <- lapply(X = S2_items_update[[i]],
-                                     FUN = function(x, offset){x - offset}, offset)
+                                     FUN = function(x, offset){x - offset},
+                                     offset)
   }
   S2_items_final <- list()
   for (dateAcq in as.character(item_collection$acquisitionDate)){
-    if (is.null(S2_items[[dateAcq]])) S2_items_final[[dateAcq]] <- S2_items_update[[dateAcq]]
-    if (is.null(S2_items_update[[dateAcq]])) S2_items_final[[dateAcq]] <- S2_items[[dateAcq]]
+    if (is.null(S2_items[[dateAcq]]))
+      S2_items_final[[dateAcq]] <- S2_items_update[[dateAcq]]
+    if (is.null(S2_items_update[[dateAcq]]))
+      S2_items_final[[dateAcq]] <- S2_items[[dateAcq]]
     if (! is.null(S2_items_update[[dateAcq]]) & ! is.null(S2_items_update[[dateAcq]]))
       S2_items_final[[dateAcq]] <- c(S2_items[[dateAcq]], S2_items_update[[dateAcq]])
     S2_items_final[[dateAcq]] <- S2_items_final[[dateAcq]][asset_names]
@@ -77,15 +73,19 @@ download_s2 <- function(aoi, raster_dir, collection_path, iChar, resolution,
                                    FUN = terra::resample,
                                    template_Rast[[1]], method = 'near')
     acq <- as.character(item_collection$acquisitionDate[[i]])
-    s2_items[[acq]] <- correct_s2stack(s2_items = S2_items_final[[i]],
-                                       acq = acq, raster_dir = raster_dir,
-                                       aoi = aoi, offset_B2 = offset_B2,
-                                       corr_BRF = corr_BRF,
-                                       bands2correct = bands2correct)
+    s2_items[[acq]] <- correct_s2_stack(s2_items = S2_items_final[[i]],
+                                        acq = acq, raster_dir = raster_dir,
+                                        aoi = aoi, offset_B2 = offset_B2,
+                                        corr_BRF = corr_BRF,
+                                        bands2correct = bands2correct)
     # save reflectance file
     if (writeoutput){
-      if (is.null(siteName)) filename <- file.path(raster_dir, paste0('plot_',iChar,'_',acq, '.tiff'))
-      if (!is.null(siteName)) filename <- file.path(raster_dir, paste0(siteName,'_',iChar,'_',acq, '.tiff'))
+      if (is.null(siteName))
+        filename <- file.path(raster_dir, paste0('plot_',iChar,'_',
+                                                 acq, '.tiff'))
+      if (!is.null(siteName))
+        filename <- file.path(raster_dir, paste0(siteName,'_',iChar,'_',
+                                                 acq, '.tiff'))
       terra::writeRaster(x = s2_items[[acq]], filename = filename, overwrite = T)
     }
   }